completed creating note feature 2.
