+++ node-notes-cli/data.json
@@ -1,9 +1,10 @@
 {
-  "nextId": 5,
+  "nextId": 6,
   "notes": {
     "1": "The event loop is how a JavaScript runtime pushes asynchronous callbacks onto the stack once the stack is cleared.",
     "2": "Prototypal inheritance is how JavaScript objects delegate behavior.",
     "3": "In JavaScript, the value of `this` is determined when a function is called; not when it is defined.",
-    "4": "A closure is formed when a function retains access to variables in its lexical scope."
+    "4": "A closure is formed when a function retains access to variables in its lexical scope.",
+    "5": "no chance this works."
   }
 }
